requirements_pointcloud.txt
# Requirements for point cloud visualization scripts
numpy>=1.20.0
opencv-python>=4.5.0
open3d>=0.15.0
tqdm>=4.60.0
matplotlib>=3.4.0
Pillow>=8.0.0
torch>=1.9.0  # Required for depth_anything_v2 